Fix /config.js endpoint
